+++ client/index.js
@@ -1,0 +1,5 @@
+import './s-alert-template.html';
+import './s-alert-template';
+
+export { sAlert } from './s-alert';
+export { default } from './s-alert';