+++ package.js
@@ -1,21 +1,23 @@
 Package.describe({
-    'summary': 'Simple and fancy notifications / alerts / errors for Meteor using Blaze Templates',
-    'version': '3.3.0',
-    'git': 'https://github.com/vlasky/meteor-blaze-s-alert.git',
-    'name': 'vlasky:blaze-s-alert'
+    name: 'vlasky:blaze-s-alert',
+    summary: 'Simple and fancy notifications / alerts / errors for Meteor using Blaze Templates',
+    version: '3.4.0',
+    git: 'https://github.com/vlasky/meteor-blaze-s-alert.git',
+    types: 'index.d.ts'
 });
 
 Package.onUse(function (api) {
     api.versionsFrom(['METEOR@2.0', 'METEOR@3.0']);
+    api.use('ecmascript');
     api.use('mongo');
     api.use('templating@1.4.0');
     api.use('underscore');
     api.use('jquery@1.11.11 || 3.0.0', 'client');
+    api.mainModule('client/index.js', 'client');
+  // TypeScript ambient declaration file
+    api.addAssets('index.d.ts', ['client', 'server']);
     api.addFiles([
-        'client/s-alert.js',
-        'client/s-alert-collection.js',
         'client/s-alert-template.html',
-        'client/s-alert-template.js',
         'client/s-alert-default.css',
         'client/s-alert-scale.css',
         'client/s-alert-slide.css',
@@ -30,12 +32,12 @@ Package.onUse(function (api) {
 
 Package.onTest(function (api) {
     api.versionsFrom(['METEOR@2.0', 'METEOR@3.0']);
+    api.use('ecmascript');
     api.use('jquery@1.11.11 || 3.0.0', 'client');
     api.use([
         'templating@1.4.0',
-        'vlasky:blaze-s-alert@3.3.0',
+        'vlasky:blaze-s-alert@3.4.0',
         'meteortesting:mocha'
     ]);
-    api.addFiles(['tests/s-alert-test.js'], 'client');
+    api.mainModule('tests/s-alert-test.js', 'client');
 });
-